refactor: moved main function to its own file
To avoid conflicts with the tests
