+++ includes/init.hpp
@@ -1,0 +1,7 @@
+#ifndef ANTIWASM_INIT_HPP
+#define ANTIWASM_INIT_HPP
+
+#include "scanner.hpp"
+#include <boost/log/trivial.hpp>
+
+#endif //ANTIWASM_INIT_HPP
