+++ src/scanner.cpp
@@ -1,22 +1,11 @@
 #include "../includes/scanner.hpp"
 
-int main(int argc, char **argv) {
-    if(argc != 2) {
-        std::cout << "Enter a filename" << std::endl;
-        return -1;
-    }
-
-    char* classFile = argv[1];
-    BOOST_LOG_TRIVIAL(debug) << "Filename: " << classFile;
-
-    antiwasm::parse(classFile);
-}
-
 namespace antiwasm {
     int parse(const char* classFile) {
         std::shared_ptr<Driver> driver = Driver::GetInstance();
         if(!driver->OpenFile(classFile)) {
             BOOST_LOG_TRIVIAL(error) << "Cannot open file";
+            return -1;
         }
 
         unsigned char *uBuffer = (unsigned char*)malloc(sizeof(unsigned char) * 4);
@@ -24,7 +13,7 @@ namespace antiwasm {
         //Magic header
         uBuffer = driver->GetNextBytes(4);
         if(antiwasm::checkMagicNumber(uBuffer) == false) {
-            exit(0);
+            return -1;
         }
 
         free(uBuffer);
@@ -33,7 +22,7 @@ namespace antiwasm {
         //Version number
         uBuffer = driver->GetNextBytes(4);
         if(antiwasm::checkVersion(uBuffer) == false) {
-            exit(0);
+            return -1;
         }
 
         free(uBuffer);
